info to format gt

+++ call_sv_2.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 #change tools to your path, notice that, you should install SURVIVOR from https://github.com/panguangze/SURVIVOR.git
-set -x
+#set -x
 SVABA=/home/panguangze/miniconda3/envs/apps/bin/svaba
-CONFIGMANTA=/home/panguangze/miniconda3/envs/apps/bin/configManta.p
-SAMTOOLS= ~/apps/usr/local/bin/samtools
-BCFTOOLS= ~/apps/usr/local/bin/bcftools
+CONFIGMANTA=/home/panguangze/miniconda3/envs/apps/bin/configManta.py
+SAMTOOLS=~/apps/usr/local/bin/samtools
+BCFTOOLS=~/apps/usr/local/bin/bcftools
 LUMPY_EXPRESS=/home/panguangze/miniconda3/envs/apps/bin/lumpyexpress
 SVTYPER=/home/panguangze/miniconda3/envs/apps/bin/svtyper-sso
 SCRIPTS=/home/panguangze/tmp_file/esv_pipe/scripts
@@ -12,7 +12,7 @@ SURVIVOR=~/tmp_file/SURVIVOR/Debug/SURVIVOR
 EXTRACT_HAIR=ExtractHAIRs
 SPECHAP=SpecHap
 ESplitReads_BwaMem=/home/panguangze/miniconda3/envs/apps/bin/extractSplitReads_BwaMem
-DELLY=/home/panguangze/miniconda3/envs/apps/bin/delly
+DELLY=/home/panguangze/tmp_file/delly/src/delly
 PYTHON3=/home/panguangze/miniconda3/envs/apps/bin/python
 PYTHON2=python2
 BGZIP=bgzip
@@ -47,35 +47,59 @@ fi
 #   $SAMTOOLS index $bam -@ $threads
 # fi
 # # svaba
-$SVABA run -t $bam -G $ref -a $out_dir/svaba/svaba --read-tracking --germline -p $threads
-$PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/svaba/svaba.svaba.sv.vcf > $out_dir/svaba/svaba.svtyper.sv.vcf
-$PYTHON2 $SCRIPTS/svaba_ano.py $out_dir/svaba/svaba.svtyper.sv.vcf > $out_dir/svaba/svaba.adjusted.vcf
+if [ ! -f $out_dir/svaba/svaba.svaba.sv.vcf ]; then
+	$SVABA run -t $bam -G $ref -a $out_dir/svaba/svaba --read-tracking --germline -p $threads
+	$PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/svaba/svaba.svaba.sv.vcf > $out_dir/svaba/svaba.svtyper.sv.vcf
+	$PYTHON2 $SCRIPTS/svaba_ano.py $out_dir/svaba/svaba.svtyper.sv.vcf > $out_dir/svaba/svaba.adjusted.vcf
+fi
+#clean
+rm $out_dir/svaba/svaba.bps.txt.gz
+rm $out_dir/svaba/svaba.contigs.bam
+rm $out_dir/svaba/svaba.alignments.txt.gz
+rm $out_dir/svaba/svaba.svaba.unfiltered.indel.vcf
+bgzip $out_dir/svaba/svaba.svaba.indel.vcf
+rm $out_dir/svaba/svaba.log
 # manta
-$CONFIGMANTA --bam $bam --referenceFasta $ref --runDir $out_dir/manta --generateEvidenceBam
-$out_dir/manta/runWorkflow.py -m local -j $threads -g 150
-gunzip $out_dir/manta/results/variants/diploidSV.vcf.gz
-cp $out_dir/manta/results/variants/diploidSV.vcf  $out_dir/manta/manta.svtyper.vcf
-time $PYTHON3 $SCRIPTS/parse.py --manta -v $out_dir/manta/manta.svtyper.vcf -b $out_dir/manta/results/evidence/evidence_0.$bam_filename -o $out_dir/manta/manta.evidence.vcf
-$PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/manta/manta.evidence.vcf > $out_dir/manta/manta.adjusted.vcf
+if [ ! -f $out_dir/manta/manta.svtyper.vcf ]; then
+	$CONFIGMANTA --bam $bam --referenceFasta $ref --runDir $out_dir/manta --generateEvidenceBam
+	$out_dir/manta/runWorkflow.py -m local -j $threads -g 150
+	gunzip $out_dir/manta/results/variants/diploidSV.vcf.gz
+	cp $out_dir/manta/results/variants/diploidSV.vcf  $out_dir/manta/manta.svtyper.vcf
+	#time $PYTHON3 $SCRIPTS/parse.py --manta -v $out_dir/manta/manta.svtyper.vcf -b $out_dir/manta/results/evidence/evidence_0.$bam_filename -o $out_dir/manta/manta.evidence.vcf
+fi
+$PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/manta/manta.svtyper.vcf > $out_dir/manta/manta.adjusted.vcf
+
+#clean
+rm $out_dir/manta/workspace -rf
+
 # $PYTHON3 $SCRIPTS/rm_cross.py -v $out_dir/manta/manta.adjusted.vcf -o $out_dir/manta/manta.adjusted2.vcf -f $ref
 # lumpy
-$SAMTOOLS view -uF 0x0002 $bam | $SAMTOOLS view -uF 0x100 - | $SAMTOOLS view -uF 0x0004 - | $SAMTOOLS view -uF 0x0008 - | $SAMTOOLS view -bF 0x0400 - | $SAMTOOLS sort - -o $out_dir/lumpy/lumpy.discordant.sort.bam
-$SAMTOOLS view -h $bam | $ESplitReads_BwaMem -i stdin | $SAMTOOLS view -Sb - | $SAMTOOLS sort - -o $out_dir/lumpy/lumpy.sr.sort.bam
-$LUMPY_EXPRESS -B $bam -S $out_dir/lumpy/lumpy.sr.sort.bam -D $out_dir/lumpy/lumpy.discordant.sort.bam -o $out_dir/lumpy/lumpy.vcf
+if [ ! -f $out_dir/lumpy/lumpy.vcf ]; then
+	if [ ! -f $out_dir/lumpy/lumpy.discordant.sort.bam ]; then
+		$SAMTOOLS view -uF 0x0002 $bam | $SAMTOOLS view -uF 0x100 - | $SAMTOOLS view -uF 0x0004 - | $SAMTOOLS view -uF 0x0008 - | $SAMTOOLS view -bF 0x0400 - | $SAMTOOLS sort - -o $out_dir/lumpy/lumpy.discordant.sort.bam
+	fi
+	$SAMTOOLS view -h $bam | $PYTHON2 $ESplitReads_BwaMem -i stdin | $SAMTOOLS view -Sb - | $SAMTOOLS sort - -o $out_dir/lumpy/lumpy.sr.sort.bam
+	$LUMPY_EXPRESS -B $bam -S $out_dir/lumpy/lumpy.sr.sort.bam -D $out_dir/lumpy/lumpy.discordant.sort.bam -o $out_dir/lumpy/lumpy.vcf
+else
+	rm $out_dir/lumpy/lumpy.discordant.sort.bam
+	rm $out_dir/lumpy/lumpy.sr.sort.bam
+fi
 $PYTHON3 $SCRIPTS/parse.py --lumpy -v $out_dir/lumpy/lumpy.vcf -o $out_dir/lumpy/lumpy.evidence.vcf
 $SVTYPER -B $bam -i $out_dir/lumpy/lumpy.evidence.vcf --core $threads -o $out_dir/lumpy/lumpy.svtyper.vcf
 $PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/lumpy/lumpy.svtyper.vcf > $out_dir/lumpy/lumpy.adj.vcf
-$PYTHON3 $SCRIPTS/parse2.py -v $out_dir/lumpy/lumpy.adj.vcf --lumpy -o $out_dir/lumpy/lumpy.adjusted.vcf #fix _2 have no reads error
+#$PYTHON3 $SCRIPTS/parse2.py -v $out_dir/lumpy/lumpy.adj.vcf --lumpy -o $out_dir/lumpy/lumpy.adjusted.vcf #fix _2 have no reads error
 # $PYTHON3 $SCRIPTS/rm_cross.py -v $out_dir/lumpy/lumpy.adjusted.vcf -o $out_dir/lumpy/lumpy.adjusted2.vcf -f $ref
 # trans to BND format 
 #$PYTHON3 $SCRIPTS/trans_to_BND_format.py -v $out_dir/lumpy/lumpy.adjusted.vcf -f $ref -o $out_dir/lumpy/lumpy.adjusted.BND.vcf
 
 
 #delly
-$DELLY call -g $ref -o $out_dir/delly/delly.bcf  $bam --dump $out_dir/delly/delly.dump.gz
-$BCFTOOLS view $out_dir/delly/delly.bcf > $out_dir/delly/delly.vcf
-cp $out_dir/delly/delly.vcf  $out_dir/delly/delly.svtyper.vcf
-gunzip $out_dir/delly/delly.dump.gz
+if [ ! -f $out_dir/delly/delly.vcf ]; then
+	$DELLY call -g $ref -o $out_dir/delly/delly.bcf  $bam --dump $out_dir/delly/delly.dump.gz
+	$BCFTOOLS view $out_dir/delly/delly.bcf > $out_dir/delly/delly.vcf
+	cp $out_dir/delly/delly.vcf  $out_dir/delly/delly.svtyper.vcf
+	gunzip $out_dir/delly/delly.dump.gz
+fi
 $BCFTOOLS view -f 'PASS' $out_dir/delly/delly.svtyper.vcf -o $out_dir/delly/delly.pass.vcf
 $PYTHON3 $SCRIPTS/pdelly.py -v $out_dir/delly/delly.pass.vcf -r $ref -o $out_dir/delly/delly.evidence.vcf -d $out_dir/delly/delly.dump
 $PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/delly/delly.evidence.vcf > $out_dir/delly/delly.adjusted.vcf
@@ -84,10 +108,10 @@ $PYTHON3 $SCRIPTS/adjust_svtyper_genotypes.py $out_dir/delly/delly.evidence.vcf 
 #$PYTHON3 $SCRIPTS/trans_to_BND_format.py -v $out_dir/delly/delly.adjusted.vcf -f $ref -o $out_dir/delly/delly.adjusted.BND.vcf
 # generate input for survivor
 
-bcftools view -f PASS -i "FORMAT/RV>=10" $out_dir/delly/delly.adjusted.vcf $out_dir/delly/delly.sr10.vcf
-bcftools view -i "FORMAT/SR + FORMAT/PE >= 10" $out_dir/lumpy/lumpy.adjusted.vcf -o $out_dir/lumpy/lumpy.sr10.vcf
-bcftools view -i "FORMAT/FT='PASS' && FORMAT/SR[0:1]>=10" $out_dir/manta/manta.adjusted.vcf -o $out_dir/manta/manta.sr10.vcf
-bcftools view -i "FORMAT/SR>=10" $out_dir/svaba/svaba.adjusted.vcf -o $out_dir/svaba/svaba.sr10.vcf
+$BCFTOOLS view -f PASS -i "FORMAT/RV>=10" $out_dir/delly/delly.adjusted.vcf $out_dir/delly/delly.sr10.vcf
+$BCFTOOLS view -i "FORMAT/SR + FORMAT/PE >= 10" $out_dir/lumpy/lumpy.adjusted.vcf -o $out_dir/lumpy/lumpy.sr10.vcf
+$BCFTOOLS view -i "FORMAT/FT='PASS' && FORMAT/SR[0:1]>=10" $out_dir/manta/manta.adjusted.vcf -o $out_dir/manta/manta.sr10.vcf
+$BCFTOOLS view -i "FORMAT/SR>=10" $out_dir/svaba/svaba.adjusted.vcf -o $out_dir/svaba/svaba.sr10.vcf
 
 
 if [ -f "$out_dir/sur.input" ]; then
